Added new activities

+++ app/src/main/java/com/example/naturalgem/PlacesRecyclerAdapter.kt
@@ -45,7 +45,9 @@ class PlacesRecyclerAdapter(val context : Context, val places : List<Place> ):
 //
 //        notifyDataSetChanged()
         val place = PlaceManager.places[position]
-        db.collection("places").document(place.documentId!!).delete()
+        if(place.documentId!=null) {
+            db.collection("places").document(place.documentId!!).delete()
+        }
 
 
     }
